cleaning up for merge into 5.1

+++ client/public/js/logout.js
@@ -1,4 +1,7 @@
-function logout() {
-    sessionStorage.clear();
-    window.location.assign('../index.html')
-}
+document.querySelector('.button-logout').addEventListener('click', () => {
+    const leaveRoom = confirm('Are you sure that you want to log off?');
+    if(leaveRoom){
+        sessionStorage.clear();
+        window.location.replace('../index.html');
+    }
+})
